Add extended-range warp filter
Introduce EXT_WARP_FILTER config parameter for cases where
fractional offsets are outside [-1,2) range from 8x8
block centre. The warped motion model is used to obtain
the centre pixel of each 4x4 block in the partition and
a translational MC is applied for the 4x4 blocks around
the projected centre using a new 6-tap warp filter.

The new filter is not enabled in this patch, we simply add
the relevant functions. The next commit will plumb this
in to the rest of the code.