New tool CONFIG_UNEVEN_4WAY: Uneven 4-way partitions
Recursive version rebased on top of
CWG-D020 Improved Signaling Scheme for ERP

Includes some relevant fixes:
- Fix av1_reset_cdf_symbol_counters and av1_avg_cdf_symbols
- Disable some assembly due to possible alignment issues.
- Optimize CDFs
- separate luma/chroma for partition_rec_cdf
- change OBMC neighbor logic
Previously, OBMC neighbor handling logic assumed that:
1. If the above block is wider than the current block, then it is the only block above current block.
2. Otherwise, the blocks above are bounded by current block's leftmost and rightmost edges
(and similar for the left neighbors)

This is no longer true with uneven 4-way partitions: It's possible that above block is (a) partially to the top-left of current block and (b) partially on top of the current block.

So, for such cases, we find overlapping region between current and above block.

Also, this overlapping region needs to be power of 2.