Refactor/Correct foreach processing for lr search
This fixes a long-standing issue in avm for search mismatch leading
to potential inefficiency in the case of multiple tiles in a frame,
or if the ru-size is smaller than the sb size.

Also includes other refactors in preparation for lr set of tools to land.

STATS_CHANGED for multitile encodes and when RU size is smaller than
SB size.