Setup bank structure for LR tools
The bank framework for LR tools is integral to the coming
CONFIG_RST_MERGECOEFFS tool. This patch introduces banks
in the existing LR tools to facilitate inclusion of that
tool subsequently. Since a bank size of 1 is equivalent to
the reference based coding framework in AV1, this change is
expected to be performance neutral.