Fixes LD with supereres
Fixes a variety of issues mostly related to AMVD searches that were
not done correctly for scaled prediction. Some of the changes in this
patch needs to be moved to main, and one to libaom as well since it
seems to be a long-standing issue inherited from libaom.
1. Makes sure aom_highbd_upsampled_pred() function handles the
case of using scaled references correctly. This was not needed
before where this function was only used for sub-pixel searches
where true references were used instead of scaled references. But
the newer AMVD searches uses this even while using scaled references.
2. Previously scaled references did not need to extend borders in
the LD case since a fast version was used. The newer AMVD searches
need extension and it is added. Also a better scaling method is
used where EXT_SUPERRES is on. The lack of proper extension causes
mismatches between multiple trials of the same resolution.
3. Fixes a long-standing issue with setting up of the buffers for
scaled reference prioir to motion searches. This seems to be there in
libaom as well and will be upstreamed to it.
4. Setup identity scale factor at the time of compressor initialization
to avoid some test failures due to uninitialized identity scale factor.