CWG-D031 RefineMV
STATS_CHANGED

This commit proposes to refine the MVs of the bi-directional compound prediction modes using bilateral matching between the prediction signals of the first and the second reference frames. The offset motion vector of a sub-block is computed by searching neighboring areas of MV0 and MV1.

A bug-fix related to BAWP is also included.