Fix TIP w/ superres when across_scale is off
The critical bug fix is the change in tip_motion_field_projection()
to use is_ref_motion_field_eligible() check.
In addition,  with any superres (1D or 2D) and TIP, the tip ref
needs to be generated in units of 8x8.
The patch also includes a few cleanups, a few compile warning fixes and
addition of a few asserts to catch problems with better diagnostics.