Signal refresh frame flags with less bits
We always signal 3 bits first. The value 1 - 7 corresponds to refresh
buffer slot 1 - 7.

When the value is 0, we read one bit to indicate having refresh or not.
If 1, we refresh buffer slot 0.
If 0, we don't refresh any buffer slot.

The scheme will not be triggered for error resilient mode.

A sequence flag is added to turn on/off this tool.

A command line option is also provided.

Change-Id: I94bd97eac996255d8cb491af1b831fa615b66771