[encoder] Don't predict global motion from low-quality refs
When deciding which frames to estimate global motion frame, skip over
any lower quality (== higher pyramid level) references. These frames
are unlikely to provide useful predictions, even if they give models
which pass the encoder's initial cost checks, because of the
availability of other high-quality references.

For frames near the bottom of the coding pyramid, the fact that we only
consider {LAST_FRAME, GOLDEN_FRAME, BWDREF_FRAME, ALTREF_FRAME} is already
enough to filter down to higher-quality refs. But for frames at the top
of the coding pyramid, there aren't many higher-quality refs available,
and so these ref slots are filled with other, lower-quality frames,
so we have to filter explicitly.

STATS_CHANGED