Skip WARP_DELTA in a case where it isn't useful
When considering WARP_DELTA, if the following conditions are true:

  * Model refinement is disabled at the syntax level, so we can
    only refine the associated MV

  * The base model is an IDENTITY model, which is the default model
    type when global motion is not in use

then the resulting prediction will be a pure translation. As we
will have already done motion search (if allowed by the block mode),
there is no reason to refine the MV further, and in fact no reason
to use WARP_DELTA over SIMPLE_TRANSLATION.

Therefore, in this case, we skip searching WARP_DELTA entirely.