Extend TX_TYPE from uint8 to uint16
Extend the data type of TX_TYPE from uint8_t to uint16_t

This is a prep work for the explicit signaling of
secondary transform set (IST).

The TX_TYPE has been used for internal representation of the
primary and secondary transform kernel type.

Then, the soon coming change in separate commit is to
explicitly signal the set of secondary transform.

This is a lossless change in terms of coding efficiency,
and should minimally affect on both encoding and decoding time.