Implement a first TEMPORAL_GLOBAL_MV based tool
Implements a first version of a tool using temporal_global_mv.
In this version, if a non-translational global motion is not
explicitly signaled, the temporal global mv for the SB is used
instead for GLOBALMV and GLOBAL_GLOBALMV modes.

STATS_CHANGED due to some fixes in the baseline related to
INTRA ref_frame values.