Fix the encoder/decoder mismatch when turning off SDP (issue #155)
In AV1, intrabc has a special check for sub 8x8 chroma cases to prevent referring to chroma pixels outside current tile. However, a sub 16x16 chroma case is introduced when CONFIG_EXT_RECUR_PARTITIONS is on, which may lead to the mismatch between encoder and decoder. This commit fix the issue by performing the check using the actual chroma size.