Add checks for signaled ref frame info (#177,#181)
Checks are added here to mark the invalid case for explicitly
signaled num_total_refs and primary_ref_frame.

The number of reference frames should be within the range of
0 < num_total_refs <= max(n_ranked, max_ref_frames).

The primary reference frame is determined at the encoder side
and signaled. It should be smaller than the implicitly derived
num_total_refs.

Change-Id: I1f41d38080150e80c6e81a4ec26383c7a2724268