ERP: Zero out cost on normatively forced partition
Currently, the encoder sometimes adds the partition cost in the rdo
loop even if the partition type is already implied by the bitstream.
This mostly happens when SDP is on for 128x128 chroma blocks.

This commit solves this by introducing a new function
`get_normative_forced_partition_type`, and zero out the partition costs
when the function detects that the current partition type is already
forced by the bitstream.