Solve the encoder assert error
When flexible RU size is enabled, the variable num_filter_classes should not be
updated in the function adjust_frame_rtype(). Instead, it should be set in the
function finalize_frame_and_unit_info().

BTW, the input parameter rsc is voided in the function
adjust_frame_rtype(), but still used inside the function, this should be
studied.