D014: Fix test_decode and testing framework and enable D014 by default.
Fixes #228 

Prior to D014 the testing framework assumed a non-null packet to be
produced for every show frame including show-existing frames.
With D014 that is no longer true, breaking many tests and the
test_decode functionality.

This patch fixes the issues by introducing an encoder side
NULL packet type folllowed by processing this packet appropriately.

With D014, everytime we call aom_codec_decode() a list of frames
is produced, starting from a non-null show frame and followed by
multiple null show-existing frames. The callbacks triggered for
the latter frames now should not call aom_codec_decode() but
instead just advance an internal pointer and return the next frame
in the output frames list.
The framework implemented incorporates the new strategy in
encode_test_driver.cc by reusing the decoder interator that was
initiated at the time of the last aom_codec_decode() call, for
subsequent null packet callbacks.

The test_decode framework uses a somewhat different API, but has
the same issue that needs to be fixed. To support this with D104
an explicit control is added to advance an internal pointer to
the output frames list instead of decoding with aom_codec_decode()
for null show-existing frames.