D014: Correct logic when PSNR packet is generated
PSNR packet should be generated for each visible frame.

So, in case of show-existing frame, packet should be generated when show-existing frame is coded and NOT when corresponding invisible frame is coded.

This is because, if we generate a packet for invisible frame, the source may have been filtered, resulting in wrong PSNR computation.

Note: This only changes the summary stats printed by the encoder; does not change the actual output bitstream.

Fixes #263