Fix mismatch debug with D014
Fixes #280.
Since with D014, we have implicit show_existing frames, the decoder
side frame pointer used in MISMATCH_DEBUG was not being advanced for
show_existing frames causing the false positives. This patch advances
the pointer for pending and imnplicit show existing frames. However
the overflow and underflow checks in MISMATCH_DEBUG needs to be
disabled for show_existing frames.