Allocate TPL recon and stats buffers conditionally
In parent version, TPL buffers were allocated in
initialize_encoder() and these buffers were not
used when TPL model is disabled. Hence, this patch
moves the allocation of TPL buffers to encoder stage
and conditionally allocates the same based on the
'enable_tpl_model' and 'lag_in_frames' flags.

Below shows the improvement to peak heap memory (in MB)
numbers for speed preset 0.

For 4k    : 5447.68 to 4082.68 (25.05%) (2 threads)
For 1080p : 2217.98 to 1866.75 (15.83%)
For 720p  : 1562.62 to 1400.83 (10.35%)
For 360p  : 1122.30 to 1075.20 (04.19%)
For 270p  : 1076.22 to 1048.57 (02.56%)