Avoid the allocation of inter_modes_info buffer for lookahead stage
In the parent version, the inter_modes_info buffer was allocated
for both lookahead and encode stages in initialize_encoder().
As this buffer is not required in lookahead stage, the buffer
allocation is moved to encode stage.

Below shows the improvement to peak heap memory (in MB) numbers
for speed preset 0.

For 4k    : 5447.68 to 5014.52 (07.95%) for 2 threads
For 1080p : 2217.98 to 1818.62 (18.00%)
For 720p  : 1562.62 to 1119.23 (28.37%)
For 360p  : 1122.30 to 680.90  (39.33%)
For 270p  : 1076.22 to 593.20  (44.88%)