FLEX_PARTITION: Bugfix for cfl_store_block() logic at boundary.
Earlier, we were only storing part of the block within frame boundary.
This works fine when `CONFIG_FLEX_PARTITION = 0` case due to the logic
in `is_partition_implied_at_boundary`.

However, when `CONFIG_FLEX_PARTITION = 1`, block size within the
boundary may not be a valid tx size. For example, block size within the
boundary maybe `24x4` for a `32x4` block at the boundary.

We fix this by always storing the full block, even if the pixels outside
the boundary may not be used later.