Fix build issues in tx_search_best_inter_candidates()
The introduction of tool 'CONFIG_COMPOUND_WARP_CAUSAL' differentiates
the way sub-block mvs are calculated for WAPRED_CAUSAL and other WARP
motion modes. In this process, the prototype change in 'assign_warpmv()'
has caused the build issues during the rebase. This CL resolves the
build issues and modified the code as per the tool
'CONFIG_COMPOUND_WARP_CAUSAL'