Optimize TokenExtra structure
In parent version, pointers in TokenExtra structure are used to
store addresses of palette related cdfs, which are consumed in
bitstream preparation stage. However, the structure size can be
reduced by storing the appropriate offsets during the encode stage
instead of pointers. Thus the required address can be derived in
the bitstream preparation stage using these offsets which can
aid in reducing the heap memory requirements of the encoder.

Resolution  Memory Savings (%)
3840x2160     22.53 (2 threads)
1920x1080     15.43 (1 thread)
1280x720       9.44 (1 thread)
640x360        6.06 (1 thread)
480x270        4.84 (1 thread)