Reduce clamp range of a[] in av1_calc_affine_autocorrelation_matrix()
The array a[] in av1_calc_affine_autocorrelation_matrix() is clamped
to [-AFFINE_CLAMP_VAL, AFFINE_CLAMP_VAL] (i.e., -2^15 to 2^15), which
is 32-bit signed range. In this CL, the upper bound of 2^15 is
modified to (2^15 -1) so that the clamped value falls in the 16-bit
signed range. Also, the SIMD module is changed to handle the
operations in 16-bit instead of 32-bit.

With this change, the module level SIMD time reduction is 5%-8%.
The scaling of the SIMD module w.r.t C is as follows.

BlockSize   Reference    Modified
            (parent)     (current)
8x8          4.99x        5.30x
8x16         5.36x        5.98x
16x8         5.46x        5.69x
16x16        5.66x        5.93x

Change-Id: If0b6b223868f034b5eb769a4258c3e203b321c5a