Fix energy calculation during CCTX angle RD evaluation
CCTX angle evaluation is conditionally skipped based on energy of
dequantized coefficients and the dequantized coefficients are stored
with 32-bit precision. In parent version, the energy was calculated
as if dequantized coefficients are stored in 16-bit buffers. This
patch corrects the energy calculation and removes one stale
comment as well.