Optimize dynamic memory alloaction in av1_optimize_txb_new()
In the parent version, the coef_info structure was dynamically
allocated in the av1_optimize_txb_new() function. As the
function is called a large number of times, the heap memory
allocation adds a significant overhead which reduces the
performance of the function.

In this patch, the structure is moved to macroblock structure
such that the overhead due to repetitive memory allocation is
avoided.