CONFIG_REFINEMV : Fix reference area pixel padding
In case of reference area padding, the pixel used to
pad the left side of the reference block is 'ref_row[0]’,
i.e., the leftmost pixel of a particular row in ref frame.
However, as the left coordinate of the defined reference
area is ‘ref_area->pad_block.x0’, the pixel used to pad
the left side should be ‘ref_row[ref_area->pad_block.x0]’
instead.

STATS_CHANGED

Change-Id: I98a36cd4b30365af30a0c103994bb1bdbe748597