Fix issues with OPFL refinement for 444/422 subsampling
The optical flow sub-block size for chroma is always set as
4x4, irrespective of the chroma subsampling. The optical
flow refined MV offsets of luma component stored in the
buffer 'xd->mv_delta' are reused for chroma components.
However, for 444/422 subsampling, for chroma, values of
'xd->mv_delta' that have not been written are accessed.

In this patch, sub-block size is modified to compute based
on subsampling such that only valid values of 'xd->mv_delta'
buffer are accessed. Also, the redundant memset of
'xd->mv_delta' is removed.

This fixes the concerns 2 and 5 of issue #335.

STATS_CHANGED for encoding path of 444 and 422 subsampling.