Fix issues when affine refinement is disabled
In case of DAMR, colocated luma warp parameters
are used for chroma components. Also, for luma
components, if the SAD between P0 and P1 is
smaller than a threshold, affine refinement is
disabled and traditional optical flow refinement
is used. In cases where affine refinement is
disabled, chroma inter prediction happens with
the default warp parameters and the default warp
parameters are used in the TMVP list population
which are incorrect.

In this patch, for chroma components, when affine
refinement for the colocated luma is disabled,
the base motion vector is added to optical flow
refined motion vectors to perform inter prediction.
Also, the TMVP list population is corrected.

This fixes issue #412.

STATS_CHANGED