Avoid redundant default initialization of warped motion parameters
The 'wms' structure in 'build_inter_predictors_8x8_and_bigger()'
stores warp motion related information during DAMR for non-refinemv
cases. When affine is enabled, this structure is populated with
motion information, which is then used for pred data calculation
and chroma refinement. This change avoids the redundant default
initialization of the 'WarpedMotionParams', as it is only
populated and utilized for the necessary sub-blocks.