encode_block(): RDO reconstruction bugfix for Inter IST
- For signaling, secondary transform type is skipped for eob <= 3 -- and
assumed to be disabled.
- However, during RDO in encode_block(), the reconstruction using
  inverse transform was still applying secondary transform on encoder side.

This led to reconstruction mismatch between encoder and decoder.
We fix this by explicitly disabling secondary transform after RDO in
encode_block(), to match what happens during signaling.

Also, we update the assertion in av1_inverse_transform_block() to avoid
this bug in future.

STATS_CHANGED