set_mi_row_col: clip last row/col to be within frame boundary.
When finding the bottom-right mi location for a chroma block covering
multiple luma units, in some cases the bottom-right location maybe
outside the frame boundary, and be a NULL pointer.

To avoid this issue, we clip the bottom-right row/col to be within the
frame boundary.

This also aligns the code with get_mi_location() in av1_loopfilter.c.