Lossless is enabled in aomenc iff tcq and ph are off
Previously lossless mode could be enabled in aomenc by one of
two options:
(1) Set --qp=<n> to be 0 for 8 bit, -48 for 10-bit and -96 for 12-bit
(2) Set --lossless=1 flag

This patch changes option (1) above so that lossless is set only if
tcq and parity hiding are both off. Further option (2) now enforces that
tcq and parity hiding are turned off in addition to setting qp to 0.