Disable tcq/ph if any segment is lossless in frame
This patch changes the behavior of tcq and ph to be allowed
only if none of the segments in a frame are lossless. If any one
segment is lossless tcq / ph cannot be used.

Prior to this patch, tcq / ph were disabled only if all segments
were coded lossless. With this patch, we can therefore now have certain
segments that are guaranteed lossless and certain segments
tht are not, which was not possible before.