Ensure a lossless segment_id cannot be switched
Solves issue #769.
A previous patch disallowed switch to a lossless segment from
a non-lossless one. That was a normative decoder behavior change.
This patch ensures that a lossless segment cannot be switched
to a non-lossless segment on the encoder side. This is needed
to make sure the loopfilter parameters stay consistent on encoder
and decoder sides. This is a non-normative change and is needed
only because of the specific implementation in AVM where segment
id switches happen only at the packing stage.