F315: Seq level max_drl_bits signaling
CONFIG_SEQ_MAX_DRL_BITS implements the new signaling mechanism.

In the new mechanism, a default value of max_drl_bits and max_bvp_drl_bits
is signaled at the sequence level along with a bit each to indicate
if these values can change from frame to frame.

If they are not allowed to change, all frames use the default one
signaled at the sequence level. If they are allowed to change, then
for each frame, a single bit is first used to indicate if the value
used should be the same as the one at the sequence level or different.
If different a new value is signaled.

STATS_CHANGED