Make tile sizes identical for intra/inter frames in case of non-uniform tiles
With non-uniform tile spacing, the size of the tile is computed
based on max SB size of the frame. Since intra frame normatively
uses SB size as 128x128 for encoder configured with max SB size
as 256x256, tile sizes of intra and inter frames do not match for
this tile configuration.

This patch modifies the tile size computation of the intra frame
to be identical with the inter frame when encoder uses max SB
size as 256x256.

This patch also extends the unit test for sb size 256x256 in case
of non-uniform tile spacing with appropriate changes for intra
frame.

STATS_CHANGED for non-uniform tile spacing