Improve LR convolve non-separable SIMD for single class
This patch adds AVX2 implementation of the 16/13-tap symmetric
luma filters for 8x8 block so that the filtering can be applied
at 8x8 block size when number of class is 1.

The average scaling number w.r.t. C for block size 8x8 is:
16-tap      19.5x
13-tap      15.3x

A timer reduction 0.7% is observed for a A5 test-set.