Optimize resample_output
The C function of resample_output is reordered
such that the vertical interpolation is SIMD friendly.

With the SIMD, data-driven intra prediction (DIP) is much faster.

For allintra mode, DIP is often used in A1 test set, with up to
8% decoding time before the change.

After the change, the percentage of DIP in decoding time significantly
drops.

As an example, cip A12, QP 135, decoding speed improves from
3.24 fps --> 3.50 fps.

Change-Id: I1aa4dab8248cf5224db35bd6ed72f057bf1d1f74