Improve ftell and fseek calls in common/obudec.c
Use the ftello and fseeko wrappers defined in common/tools_common.h so
that we get large file support (file sizes > 2GB). ftell and fseek use
the `long` type for file offsets, which are limited to < 2GB file sizes
if `long` is 32 bits.

Check the return value of fseeko because it may fail.

Note: This merge request does NOT deal with the issue that POSIX allows
fseek() to seek beyond the end of file. This means we cannot detect an
OBU size larger than the amount of data in the file.