Enable MT for LR frame module
In the parent version, MT implementation of LR was done for each
restoration unit in a row where the height of each job was LRU size
causing more thread wastage.
The current design modifies the implementation where the height of
each job is stripe height. The MT scaling for LR is ~4.3x for A2
testset with --threads=8.
Also, LR MT is enabled for disable_loopfilters_across_tiles = 0 case.

Change-Id: Ia85de1653f12c27703842433c74d2fb49a473e16