Prepare luma buffer of cross-component wienerns when required
In the parent code, the downsample luma buffer required for
cross-component wienerns is prepared even when the frame
restoration type is RESTORE_NONE for chroma planes.
This patch modifies the function call conditionally such that
the preparation of the luma buffer happens only when frame
restoration is enabled for chroma planes.

Change-Id: I9c452f374d9fd586c70199ac6ea87a971f71aeac