Clean up the variable `pv_skip_txfm` in deblocking
In set_lpf_parameters() function, setting of
`pv_skip_txfm` used `plane_type`and `tree_type` of
the current block instead of previous block. However,
it is observed that as deblocking is always tested
across prediction block boundaries the value of
`pv_skip_txfm` is redundant and can be removed.