Remove new_qmobu_added from the AV1_COMP struct
The new_qmobu_added field is intended to control whether
total_signalled_qmobu_count should be incremented. However, it is not
clear when we should reset new_qmobu_added to 0 and when we should check
if new_qmobu_added is equal to 1.

Also, new_qmobu_added appears to compensate for (i.e., undo) the
decrement of total_signalled_qmobu_count after the write_qm_obu() call
in av1_pack_bitstream(). If we decrement total_signalled_qmobu_count
without first freeing the quantizer_matrix memory in the corresponding
qm_obu struct, we may leak memory. Since that decrement is risky and
should be removed, we don't need to use new_qmobu_added to compensate
for that decrement.